requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-Levenshtein==0.23.0
networkx==3.2.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
pyyaml==6.0.1
gunicorn==21.2.0